requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# PostgreSQL (será usado depois)
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0